+++ packages/bezier-codemod/src/transforms/tests/enum-member-to-string-literal/fixtures/input1.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { ProgressBar, ProgressBarSize, ProgressBarVariant } from '@channel.io/bezier-react'
+
+export default function UploadProgress({
+  uploadProgressPercentage,
+}: {
+  uploadProgressPercentage: number
+}) {
+  return (
+    <ProgressBar
+      width='100%'
+      size={ProgressBarSize.M}
+      variant={ProgressBarVariant.GreenAlt}
+      value={uploadProgressPercentage / 100}
+    />
+  )
+}