+++ packages/bezier-icons/scripts/generate-icon-files.js
@@ -1,33 +1,8 @@
-const fs = require('fs')
 const path = require('path')
 
-const bezierIconsDirectory = path.resolve(__dirname, '..')
-const iconsJson = path.resolve(bezierIconsDirectory, 'icons.json')
-const iconsDir = path.join(bezierIconsDirectory, 'icons')
-const svgByName = JSON.parse(fs.readFileSync(iconsJson, 'utf-8'))
+const { generateIconFiles } = require('./utils/generateIconFilesFromJson')
 
-const flushAndMakeIconsDirectory = () => {
-  if (fs.existsSync(iconsDir)) {
-    fs.rmSync(iconsDir, { recursive: true, force: true })
-    fs.mkdirSync(iconsDir)
-  }
-}
+const bezierIconsDirectory = path.resolve(__dirname, '../icons')
+const iconsJson = path.resolve(__dirname, '../icons.json')
 
-const makeSvgFiles = ([iconName, svgObject]) => {
-  const svgPath = path.resolve(iconsDir, `${iconName}.svg`)
-  const { svg } = svgObject
-
-  fs.writeFileSync(svgPath, svg, 'utf-8')
-}
-
-const generateSVGFilesFromMap = () => {
-  Object.entries(svgByName)
-    .forEach(makeSvgFiles)
-}
-
-const generateIconFiles = () => {
-  flushAndMakeIconsDirectory()
-  generateSVGFilesFromMap()
-}
-
-generateIconFiles()
+generateIconFiles(iconsJson, bezierIconsDirectory)